Add enterkeyhint="enter"
(React enterKeyHint = DOM enterkeyhint)

Previous behavior:
  1. When the CreatableSelect is focused, the enter key displays a 'next'
     icon
  2. Pressing the enter key advances to the next input field

New behavior, which is what I want:
  3. When the CreatableSelect is focused, the enter key displays an 'enter'
     icon
  4. Pressing the enter key selects the first dropdown item (the user can
     type to filter, which changes what the first dropdown item is)

This change only explicitly accomplishes #3. But it also apparently
accomplishes #4 on my phone.

This is surprising! enterkeyhint isn't for changing behavior, just the
appearance. (Also, even if I use enterkeyhint="next", we still get #4 instead
of #2.)

I'm guessing the reason it *does* change behavior here is that mobile Chrome's
behavior is to infer that (A) the presence of a next input field means the
enter key should advance to the next input, but also (B) the presence of the
enterkeyhint attribute (with any value) means to Chrome that I have my own
plans so it disregards A.